pythonanywhere_deploy/requirements.txt
google-search-results==2.4.2
requests==2.31.0
beautifulsoup4==4.12.3
jinja2==3.1.3
python-dotenv==1.0.1
lxml==5.1.0
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
libsql-experimental==0.0.34
